msidb: Add support for exporting with short (DOS) filenames.
msidb's "-s" flag modifies the database export to use short (DOS)
filenames instead of full table names.  This flag is convenient
because it uses the same filenames that the import (-i) option
expects.  For example, the InstallExecuteSequence table gets imported
(or exported with this flag) as InstallE.idt where the normal export
option uses InstallExecuteSequence.idt.

Signed-off-by: Erich E. Hoover <erich.e.hoover@gmail.com>
Signed-off-by: Hans Leidekker <hans@codeweavers.com>
Signed-off-by: Alexandre Julliard <julliard@winehq.org>