msidb: Add support for wildcard table import.
msidb allows multiple tables to be imported with the use of
standard wildcards.  For example, both Feature and
FeatureComponent tables can be imported simultaneously like so:
msidb -d package.msi -f . -i 'Feature*'

Please note that it is important to quote the wildcard to prevent
files from being passed instead of the intended pattern.

Signed-off-by: Erich E. Hoover <erich.e.hoover@gmail.com>
Signed-off-by: Hans Leidekker <hans@codeweavers.com>
Signed-off-by: Alexandre Julliard <julliard@winehq.org>