avicap32: Verify v4l devices can capture before returning them.
/dev/video* device nodes aren't all capture devices, and returning
those that aren't results in devenum reporting them to
applications, which will later fail when opening them with
IMoniker::BindToObject().

avicap32 already tries to call VIDIOC_QUERYCAP to check whether
it's dealing with a v4l device, but it does not check a successful
result any further. Get it to verify the device supports the
V4L2_CAP_VIDEO_CAPTURE flag, like we do in qcap/v4l.c.

Signed-off-by: Damjan Jovanovic <damjan.jov@gmail.com>
Signed-off-by: Alexandre Julliard <julliard@winehq.org>