wined3d: Allow loading buffers mapped through system memory.
Fixes rendering dirt on the wind shield and GL_INVALID_OPERATION
errors in "Need For Speed Shift 2".

The test succeeds on Windows with retail Direct3D 9, but fails
with debug Direct3D 9 if selected in DirectX SDK control panel:
error code is returned from _DrawIndexedPrimitive() and the
triangle is not rendered.

Signed-off-by: Paul Gofman <gofmanp@gmail.com>
Signed-off-by: Henri Verbeet <hverbeet@codeweavers.com>
Signed-off-by: Alexandre Julliard <julliard@winehq.org>