strmbase: Get rid of the base transform.
This is providing very little of value. The code in Receive() and in the state
management methods should probably be moved to the base pin and filter
respectively; the media type management is just wrong (and our two extant
transforms should not handle it the same way anyway), and most of the rest just
passes through to the transform's own callback table.

Signed-off-by: Zebediah Figura <zfigura@codeweavers.com>
Signed-off-by: Alexandre Julliard <julliard@winehq.org>