shell32/tests: Fix the temporary directory creation in shlexec.
Since the test no longer uses a random directory name, the directory is
more likely to exist already. So use it even if it already exists.
Also on some Windows versions (e.g. Windos 2008) GetLongPathName() does
not return the long name if the path does not exist. So do the
conversion before appending the new directory name.

Signed-off-by: Francois Gouget <fgouget@codeweavers.com>
Signed-off-by: Alexandre Julliard <julliard@winehq.org>